+++ src/org/nulllab/nullengine/core/geometry/Point2D.java
@@ -2,27 +2,27 @@ package org.nulllab.nullengine.core.geometry;
 
 public class Point2D {
 
-  protected int x;
-  protected int y;
+  protected double x;
+  protected double y;
 
-  public Point2D(int x, int y) {
+  public Point2D(double x, double y) {
     this.x = x;
     this.y = y;
   }
 
-  public int getX() {
+  public double getX() {
     return this.x;
   }
 
-  public void setX(int x) {
+  public void setX(double x) {
     this.x = x;
   }
 
-  public int getY() {
+  public double getY() {
     return this.y;
   }
 
-  public void setY(int y) {
+  public void setY(double y) {
     this.y = y;
   }
