+++ src/org/nulllab/nullengine/core/graphics/Drawable.java
@@ -4,7 +4,7 @@ import org.nulllab.nullengine.core.geometry.Point2D;
 
 public interface Drawable<Graphics extends Canvas> {
 
-  public void draw(Graphics graphics, int x, int y);
+  public void draw(Graphics graphics, double x, double y);
 
   default void draw(Graphics graphics, Point2D point) {
     draw(graphics, point.getX(), point.getX());
